bug fixes and improvement
