+++ src/note_backend/src/note_backend/src/store.rs
@@ -1,0 +1,36 @@
+use ic_stable_structures::memory_manager::{MemoryId, MemoryManager, VirtualMemory};
+use ic_stable_structures::{BoundedStorable, Cell, DefaultMemoryImpl, StableBTreeMap, Storable};
+use std::{borrow::Cow, cell::RefCell};
+
+use crate::types::Note;
+
+// Define Struct and types
+type Memory = VirtualMemory<DefaultMemoryImpl>;
+type IdCell = Cell<u64, Memory>;
+
+// Define Memory manager and ID counter
+thread_local! {
+    static MEMORY_MANAGER: RefCell<MemoryManager<DefaultMemoryImpl>> = RefCell::new(
+        MemoryManager::init(DefaultMemoryImpl::default())
+    );
+
+   pub static ID_COUNTER: RefCell<IdCell> = RefCell::new(
+        IdCell::init(MEMORY_MANAGER.with(|m| m.borrow().get(MemoryId::new(0))), 0)
+            .expect("Cannot create a counter")
+    );
+
+   pub static NOTES_STORAGE: RefCell<StableBTreeMap<u64, Note, Memory>> =
+        RefCell::new(StableBTreeMap::init(
+            MEMORY_MANAGER.with(|m| m.borrow().get(MemoryId::new(1)))
+    ));
+}
+
+// helper function to add note to storage
+pub fn do_insert(note: &Note) {
+    NOTES_STORAGE.with(|storage| storage.borrow_mut().insert(note.id, note.clone()));
+}
+
+// helper function to retrieve Note by ID
+pub fn _get_note(id: &u64) -> Option<Note> {
+    NOTES_STORAGE.with(|storage| storage.borrow().get(id))
+}
